backend/expenses/urls.py
from django.urls import path, include
from rest_framework.routers import DefaultRouter
from rest_framework_simplejwt.views import TokenObtainPairView, TokenRefreshView
from .views import (
    ExpenseViewSet, CategoryViewSet, BudgetViewSet, 
    AnalyticsSummaryView, RegisterView, EmailWebhookView
)

router = DefaultRouter()
router.register(r'expenses', ExpenseViewSet, basename='expense')
router.register(r'categories', CategoryViewSet, basename='category')
router.register(r'budgets', BudgetViewSet, basename='budget')

urlpatterns = [
    path('', include(router.urls)),
    path('auth/register/', RegisterView.as_view(), name='auth_register'),
    path('auth/login/', TokenObtainPairView.as_view(), name='token_obtain_pair'),
    path('auth/refresh/', TokenRefreshView.as_view(), name='token_refresh'),
    path('analytics/summary/', AnalyticsSummaryView.as_view(), name='analytics_summary'),
    path('webhooks/email/', EmailWebhookView.as_view(), name='email_webhook'),
]
