backend/requirements.txt
django==5.0.3
djangorestframework==3.15.1
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.1
celery==5.3.6
redis==5.0.3
psycopg2-binary==2.9.9
python-dotenv==1.0.1
